services/performance_prediction/pyproject.toml
[project]
name = "performance-prediction"
version = "0.1.0"
description = "Machine learning flow with sagemaker"
authors = [
    {name = "Caephas",email = "arinzeobidiegwu@gmail.com"}
]
requires-python = ">=3.9,<3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "matplotlib (>=3.9,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "joblib (>=1.4.2,<2.0.0)",
    "sagemaker (>=2.243.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
